refactor: decouple privileged sessions
