+++ packages/webapp/pages/account/security.tsx
@@ -13,6 +13,16 @@ import React, {
   useRef,
   useState,
 } from 'react';
+import { useMutation, useQuery } from 'react-query';
+import {
+  AuthFlow,
+  initializeKratosFlow,
+  submitKratosFlow,
+} from '@dailydotdev/shared/src/lib/kratos';
+import {
+  getNodeValue,
+  ValidateResetPassword,
+} from '@dailydotdev/shared/src/lib/auth';
 import { AccountSecurityDisplay as Display } from '../../components/layouts/AccountLayout/common';
 import { getAccountLayout } from '../../components/layouts/AccountLayout';
 import AccountSecurityDefault from '../../components/layouts/AccountLayout/Security';
@@ -20,6 +30,10 @@ import EmailFormPage from '../../components/layouts/AccountLayout/Security/Email
 import ChangePasswordForm, {
   ChangePasswordParams,
 } from '../../components/layouts/AccountLayout/Security/ChangePasswordForm';
+
+interface ResetFormParams {
+  password: string;
+}
 
 const AccountSecurityPage = (): ReactElement => {
   const { user } = useContext(AuthContext);
@@ -31,20 +45,39 @@ const AccountSecurityPage = (): ReactElement => {
     displayToast('Password reset successful!');
     setActiveDisplay(Display.Default);
   };
+  const { data: settings } = useQuery(['settings'], () =>
+    initializeKratosFlow(AuthFlow.Settings),
+  );
 
   const {
     showVerifySession,
-    onUpdatePassword,
+    initializePrivilegedSession,
     onPasswordLogin,
     onCloseVerifySession,
     onSocialLogin,
-  } = usePrivilegedSession({
-    onResetSuccessful: onPasswordReset,
-    onSessionVerified: () => {
-      const form = formToJson<ChangePasswordParams>(formRef.current);
-      onUpdatePassword(form);
+  } = usePrivilegedSession();
+
+  const { mutateAsync: resetPassword } = useMutation(
+    (params: ValidateResetPassword) => submitKratosFlow(params),
+    {
+      onSuccess: async ({ redirect }) => {
+        if (redirect) {
+          return initializePrivilegedSession(redirect);
+        }
+
+        return onPasswordReset?.();
+      },
     },
-  });
+  );
+
+  const onUpdatePassword = ({ password }: ResetFormParams) => {
+    const { action, nodes } = settings.ui;
+    const csrfToken = getNodeValue('csrf_token', nodes);
+    resetPassword({
+      action,
+      params: { csrf_token: csrfToken, method: 'password', password },
+    });
+  };
 
   if (!user) {
     return null;
